+++ mail/mail-common/src/db/migrations/v006_proton_mail_message_tables.sql
@@ -103,6 +103,7 @@ CREATE TABLE draft_attachment_metadata (
   action_id INTEGER DEFAULT NULL,
   display_order INTEGER NOT NULL DEFAULT 0,
   ownership INTEGER NOT NULL,
+  deleted INTEGER NOT NULL DEFAULT 0,
   CONSTRAINT draft_attachment_metadata_attachment_id FOREIGN KEY (local_attachment_id) REFERENCES attachments (local_id) ON DELETE CASCADE CONSTRAINT draft_attachment_metadata_metadata_id FOREIGN KEY (metadata_id) REFERENCES draft_metadata (id) ON DELETE CASCADE CONSTRAINT draft_attachment_metadata_action_id FOREIGN KEY (action_id) REFERENCES action_queue (id) ON DELETE SET NULL
 );